pyproject.toml
[tool.poetry]
name = "f1-demo"
version = "0.1.0"
description = "F1 demo"
authors = ["Michael Brewer <michael.brewer@gyft.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.17.44"
aiohttp = {extras = ["speedups"], version = "^3.7.4"}
aws-lambda-powertools = "^1.13.0"

[tool.poetry.dev-dependencies]
moto = {extras = ["all"], version = "^2.0.4"}
pytest = "^6.2.3"
pytest-asyncio = "^0.14.0"
rich = "^10.1.0"
vcrpy = "^4.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
